Delete sentence from no-warranty section.
The sentence "You may have additional rights under applicable law
which this License cannot change." is deleted. This was adapted from
no-warranty provisions in certain Microsoft licenses. The deletion is
justified by the fact that it merely states and at most provides
notice of a legal truism. If there is some significant legal benefit
to doing so, we can restore it.