"parts" is an unnecessary word here.
I don't see a specific difference between "all parts of the Covered Work"
and "all of the Covered Work".  Since Covered Work is a defined term, if
anything, "parts" should be included in its definition.  However, I tend
to believe it just isn't necessary.  "all of the Covered Work" means "all
parts".  How could it not?