Require Corresponding Source to identify, name and describe any needed Separate Works.
It's certainly historical GPL policy that Separate Works (if we can get an
adequate definition of Separate Works) need not be included as part of
Corresponding Source.

However, it's industry practice and is believed by most to be required
under GPLv2 to at least name, identify, and describe Separate Works that
are needed.

For example, for decades, the expected requirement under GPLv2 for
Microsoft Windows software has been to describe the precise proprietary
compiler suite (and its specific version number) needed to build the
software.

This change implements that long-standing policy and interpretation into
clear license text for copyleft-next.