Modify sec. 12: "the Received Work" -> "We".
This change is intended to clarify that in any instance of licensing
permission to 'relicense' under later versions is coming from the
relevant licensor.

I don't think this change would break anything, because, given a
Received Work, all the upstream 'We's' need not to explicitly remove
permission to relicense in order for 'you' to have such permission.