Modify Corresponding Source definition, mainly to delete 'install and run'.
See:
https://lists.fedorahosted.org/pipermail/copyleft-next/2012-November/000276.html
The concern expressed here is that "install and run" could be read as
an anti-lockdown requirement. Note that those verbs are used in
GPLv3's Corresponding Source definition, yet GPLv3 explicitly limits
the Installation Information requirements to 'User Products'.

Perhaps there is an alternative change that would keep 'install and
run' but add the clarification that this does not imply a GPLv3-style
anti-lockdown Installation Information requirement.