Add grace period to proprietary relicensing poison pill.
The concern here is for the company that begins with a proprietary
product and decides to release it as free software under
copyleft-next. It is likely that in such cases the company would have
triggered the earlier version of this provision, as it would have
continued to offer the proprietary version.