Fix bug introduced in Proprietary Relicensing; some rephrasing.
The immediately preceding version would have activated the trigger for
Distribution under the current version of copyleft-next.

The awkwardness of "later versions of copyleft-next" (when at the
outset "this License" is indicated to be a synonym of copyleft-next)
is noted. Similar problems exist in the GNU GPL, including GPLv3
despite the introduction in that license of a definition aimed at
eliminating the issue.