Revise Corresponding Source definition.
The main change here is the deletion of the reference to "interface
definition files", introduced in GNU GPLv2. See the discussion in the
thread "Fwd: Re: Modify Corresponding Source definition, mainly to
delete 'install and run'. (14249c2)" on the copyleft-mailing list,
e.g., this helpful posting by Ted Ts'o:
https://lists.fedorahosted.org/pipermail/copyleft-next/2012-December/000344.html

Whatever was meant to be captured by adding in the "interface
definition files" reference is, I believe, adequately addressed by
category (ii) of the Corresponding Source definition as changed by
this commit.