Minor change to README.md.
This is a further refinement to some language bkuhn had added in
recently. The main change is deleting "and contributors" from the
phrase "developers and contributors". To me this suggested that
contributors [of code] are not developers, which does not seem
right. If the intention was to sweep in non-developer contributors,
however, or if it is important to specifically mention such
contributors here, we can improve on this.