Eliminate non-year-defined definition of Minimum Period.
This change returns the 'Product Distribution' section a bit closer to
GPLv2, as the deleted language originates in a change introduced in
GPLv3. My admittedly-vague justification for this change is that, for
shipment of a given Product, there could be significant variance in
the offering of support or spare parts across Product
instances. Consider restoring this after further analysis.