Broaden Proprietary Relicensing section to cover "copyleft trolling".
The Proprietary Relicensing provision should also apply in
circumstances where 'We' (perhaps including licensors downstream from
the original licensor) offer to forgive past or future noncompliance
with respect to a work 'We' do not distribute (e.g., what would be a
downstream Derived Work but which is created by someone lateral to
'you') through purchase of a proprietary license.

There is some evidence that some purported GPL licensors are engaging
in such behavior (rather than the more familiar case of announcing at
the outset that proprietary licenses are available for the
GPL-licensed work distributed by the licensor, or some enhanced
version of it).