Date-qualify OSI-approved license set in Proprietary Relicensing provision.
The date 1 January 2013 is arbitrary (another option might be to use
the date of publication of a given released version of
copyleft-next). The point is to fix the set of OSI-approved licenses
to guard against the possibility of pathological future OSI license
approval decisions (however unlikely or remote that may be).