Make licenses "perpetual"; restore "and conditions".
The GPLv3-derived language "for the duration of ... copyright" is more
succinctly expressed by the standard "perpetual"; it is appropriate to
specify this for the patent license as well as the copyright
license. I made "terms" "terms and conditions" (which I believe was in
a much earlier revision) as a bit of drafting paranoia. The
termination provision might not be a "condition" as such, but there
may be some value in mentioning "conditions" explicitly to make
sufficiently clear that irrevocability assumes compliance with the
license's conditions.