Remove carveout from Mere Aggregation definition.
This was a descendant of the "Aggregate" definition in the final
paragraph of GPLv3 section 5. I do not believe it is necessary.

The effect of the GPLv3 'Aggregate' definition carveout is that in
cases where copyleft scope is cut off at the Aggregate boundary, the
cut-off is removed if one is conveying an aggregation under a
compilation license that "limit[s] the access or legal rights of the
compilation's users beyond what the individual works permit". As I
believe I noted in an earlier commit, this arguably goes further than
is necessary. If one supposes a copyrightable compilation C containing
works A and B, where A is copyleft-next and B is under some other
license (free or proprietary), and C meets the definition of 'Mere
Aggregation' (without the carveout), and C is governed by a license
covering the components of the compilation under terms more
restrictive than either GPLv3 or the B license, we shouldn't care (as
a policy matter for purposes of what the license of *A* should ) that
use of B is being restricted further than it was under the B
license. We should care about the effects of the C license on A. But
that already violates copyleft-next section 7, by imposing further
restrictions.

I don't see what the point is of expressing the policy in terms of
copyleft scope. Under the version of the Mere Aggregation definition
existing prior to this commit, if the idea were that the restrictive
nature of the C license justifies a broadening of copyleft scope, the
effect is that there is a violation of the copyleft requirement to
license Derived Works under copyleft-next. But we already achieve that
by the prohibition against imposition of further restrictions. The
only casualty of this change seems to be the case of the
copyleft-next-compatible subset of Bs, which could become part of a
broadened copyleft scope free of any free software license clash. But
copyleft-next does not have as its mission to save non-copyleft-next
free software from the intentionally-drafted features of their
different licenses.

Revisit if I'm missing something here, or misunderstanding the point
of GPLv3 section 5 last paragraph.