Add definition of "You".
The previous omission of a definition of "you" was deliberate (it was
one of the things cut out in the process of paring down GPLv3). Now
that copyleft-next is much smaller, there is arguably some value to
having a definition that makes an effort to unify a licensee with its
affiliates in some standard manner (perhaps mainly for the
patent-related provisions of the license).
Cf. https://www.gnu.org/licenses/gpl-faq.html#ApacheLegalEntity (I
consider the hint in this FAQ answer that such definitions are
belt-and-suspenders drafting to be fair).

The definition here is closely based on that of MPL 2.0.