Modify definition of 'Distribute'.
The clarification about intra-organizational transfers not being
distribution is removed as unnecessary. It is interesting to note
that, in my experience, a number of corporate lawyers have expressed
concern about GPL source code requirements being triggered by
intra-corporation transfers, and I suppose that was the purpose of the
deleted language. I am not even sure a 'Distribute' definition is
worth having (its inclusion may be seen as a legacy of GPLv3
"convey"), though I am keeping it for now, and I also add lowercase-d
distribute to the definition.