Limit section 10 to GPL compatibility.
It is not particularly useful or necessary to provide for outbound
AGPLv3+ compatibility (if we assume that future versions of the GPL
and AGPL will continue to be compatible to at least the same extent
GPLv3 and AGPLv3 are, an assumption that seems reasonable to make).

Since copyleft-next is outbound GPLv3 compatible, it is by definition
outbound AGPLv3 compatible. This allows for simplification of this
provision. Effective AGPLv3 compatibility can be noted in a FAQ.

Any value in specifically noting AGPL compatibility seems to be purely
political in nature. It is already a basic principle of copyleft-next
drafting that mere political benefit is never a good enough reason for
retention or inclusion of a license provision.

But that political value seems rather negligible in any case. AGPLv3
sees little use, most of it by proprietary relicensing companies or
entrepreneurs engaging in a licensing tactic that copyleft-next is
designed to discourage. By contrast, GPLv3 is a very widely used
license.