Incorporate outbound GPL compatibility into basic copyleft provision.
This commit deletes the 'Outbound GPL Compatibility' section and adds
a new paragraph to section 5, influenced by MPL 2.0's section 3.3.

With this change, outbound GPL compatibility is expressed as an
optional supplementary act that can be done when satisfying the
copyleft requirement. Distributed Derived Works must be licensed under
copyleft-next, but 'You' can *additionally* license the Derived Work
under any version of the GPL.

One thing that has changed here is that there is no explicit
permission to 'relicense' the *Received Work* (what You received,
unmodified, at least conceptually) under the GPL. It does not seem to
be necessary (much as the drafters of MPL 2.0 did not seem to regard
it as necessary to achieve downstream GPL/LGPL/AGPL compatibility
outside of the context of what the MPL calls 'Larger Works'). In any
case where outbound GPL compatibility is thought of as desirable or
necessary, we should be within the definition of 'Derived Work'.