'We' -> 'I'
Use of 'We' to refer to a single-instance upstream licensor may be
more confusing than simply using singular 'I'. The only real
disadvantage to 'I' is that it seems out of place when talking about a
corporate licensor (in common US English, at least, corporations
typically refer to themselves using 'We').