Ease comprehensibility of proprietary relicensing poison pill.
One of the problems of drafting this provision is how to define
'proprietary license'. I do not know a good way to do so in positive
terms, so I've done so by trying to say what a proprietary license
*isn't*, with the expectation that this won't sweep in anything that
shouldn't be swept in. In the previous version, a proprietary license
effectively meant a license that wasn't (i) an authorized version of
copyleft-next, (ii) an OSI-approved license (as of the beginning of
2013), or (iii) a version of the GPL.

In this commit, I replace all that with the notion of compliance with
the Open Source Definition. (I don't think we can count on OSI
approval of any particular version of copyleft-next, let alone *all*
of them, so using "OSI-approved licenses" would require restoring the
separate reference to authorized versions of copyleft-next. That
actually may be worthwhile, despite the additional length.

Another issue with this provision is that the basic hack is a bit
non-obvious to the casual reader. I've addressed this already by
changing the section heading, but here I go a step further. Instead of
saying the immediately cryptic 'sections 4 through 8 no longer apply'
(as the casual reader may not have read those sections yet), in this
new version the copyleft-next verson nominally remains but the
licensor additionally licenses the Received Work under the Apache
License 2.0.

The reference to an external license instead of the technique of
making copyleft conditions of copyleft-next disappear represents a
significant loss of elegance, but the result is hopefully easier to
understand.