'Received Work' -> 'My Work'.
"Received Work" is a suboptimal symbol for the underlying concept,
because what's being licensed is essentially what GPLv3 (using a term
from MPL 1.1) calls the 'contributor version' in its patent license
grant. It may or may not be identical to what is *actually* received
by "You", the downstream licensee.

This change may create or expose a further suboptimality, which is
that 'My Work' sounds like it might just cover "the things I added",
which is not the intention. A clarification to the definition of "My
Work" may be useful.