Remove explicit requirement for CS to be under 'this License'.
This seems to cause more potential for conceptual puzzlement than is
necessary.

Source code in the strictest sense (cf. GPLv3 definition of complete
source code) will need to comply with the conditions stated earlier in
the license anyway.

As for such things as build scripts, those that I have encountered
(and it is notable that they seem to be customarily treated as what
copyleft-next calls 'Separate Works') are of rather dubious
copyrightability. This will be all the more true of 'instructions and
similar information' and the new 'list of Separate Works'.