Add third-party beneficiary clause to object code distribution section.
This commit may be the most iconoclastic in the history of the
drafting of copyleft-next to date.

The intent of this change is that recipients of binaries, in addition
to the upstream copyright holder(s), will be able to enforce
compliance with the Corresponding Source requirements stated in the
section, if applicable law allows such a cause of action.