Generalize copyleft nullification; define 'Give' Corresponding Source.
The 'poison pill' section is now titled "Copyleft Nullification" and
adds a second trigger: if the licensor provides only a binary without
providing Corresponding Source (as has sometimes been known to happen
with the initial licensor in GPL experience). Whether this is really
worth adding (at the cost of three lines and ~16 words) is open to
question.

As a consequence of this change, the notion of providing Corresponding
Source was factored out of section 7 into a new definition of 'Give'
[Corresponding Source] in the final section, which is generalized to
cover any Distributor, including the initial licensor.