Effectively revert copyleft nullification generalization and 'Give' definition.
Trying to cover the scenario of the licensor who nominally licenses a
binary under the GPL without providing Corresponding Source seems more
trouble than it is worth.

The main point I suppose is that the licensee should be relieved from
having to comply with the (effectively impossible) Corresponding
Source requirement. However, it is possible that even worrying about
explicitly covering this scenario (though it has occurred in real
life) is falling into GPL hyperlegalization trap.