Remove 'strong vs. weak' explanation; replace 'strong' with 'non-weak'.
Looking at the direction in which copyleft-next is going, I have
recently wondered whether characterizing it as a strong copyleft
license in a specific FSF/RMS-esque sense is still entirely
accurate. A couple of provisions in 0.2.0, among the more interesting
ones in the license, provide for the disappearance of copyleft under
certain circmstances (proprietary relicensing after a grace period,
and passage of 15 years from first distribution).

I also believe the copyleft-next notion of copyleft scope may not be
exactly the same as that espoused by the FSF in its interpretation of
the GPL. It will still be the case that anyone is on safe interpretive
ground if they treat copyleft-next scope as equivalent to GPL copyleft
scope.

That said, copyleft-next shall remain a 'stronger' copyleft than the
weakest ones (like the MPL), hence my replacement of 'strong' with
'non-weak'.