Improvements to child-of-poison-pill (COPP) section.
Changes in this commit:

1) The one year grace period is eliminated. The original justification
for this was the corporate asset acquisition scenario. For
copyleft-next it seems silly to treat this as a concern. The
supposition would be that some corporation would acquire some
proprietary software product and desire to 'open-source' it under
copyleft-next, but would suffer from immediate impact of the COPP if
there is any overlap in proprietary and copyleft-next licensing (which
typically there would be). If, in a hundred years, this sort of
scenario becomes real, we or our successor license stewards can add
the grace period back in.

Perhaps there is some additional justification for the grace period,
but I do not see it at the moment.

2) COPP is now triggered only if the proprietary license is "for a
fee". This revives an earlier variant of this provision which spoke of
commercial licensing. This limitation may address some hypotheticals
critics have raised on identi.ca and in the comments to Martin
Michlmayr's LWN.net article on my FOSDEM talk.

3) In response to some suggestions from others on the mailing list, I
have replaced 'failure to satisfy the OSD' with an older formula: an
OSI-approved license, an FSF 'free' license, or a numbered version of
copyleft-next, as the negative definition of 'proprietary'.

4) In response to suggestions from others on the mailing list, I have
gone back to the more elegant solution of collapsing copyleft-next
into 'permissive-next' rather than reference a different external
license (Apache 2.0).