Inbound license compatibility for non-GNU Rule-of-Two licenses.
This change essentially implements something similar in spirit to the
suggestion, made in this GitHub issue filed by clopez:
https://github.com/richardfontana/copyleft-next/issues/15.

It may be seen as a generalization of the earlier clause in the
no-further-restrictions section allowing for inbound Apache-2.0
license compatibility; that clause is now deleted.

The carveout of the GPL and AGPLv3 is, or appears, awkward. One
possible simplification would be to add AGPLv3 as an
outbound-compatible license through the mechanism of section 3, second
paragraph.

This is actually a significant change; it practically removes
copyleft-next from the intricate license compatibility doctrine
universe associated with orthodox GPL interpretation.

I considered adding an explicit statement that Object Code Covered
Works need to be fully compliant with section 5. That perhaps can be
adequately addressed in a FAQ. The effect of this change is
reminiscent of some of the GNU-style linking exceptions that condition
the exception on providing the source code for the linked
GPL-incompatible object code.