Factor definitions of 'FSF-Free', 'OSI-Approved', (expanded) 'GPL'.
A substantive change here is that 'GPL' now means the GNU AGPL as well
as the GNU GPL, so the MPL 2.0-like compatibility mechanism of section
3, paragraph 2 now allows for AGPL compatibility directly. This is
done principally to reduce the verbosity that was previously
introduced into following section.

There is arguably some redundancy between the more specific section 3,
paragraph 2 and the more general section 4, paragraph 2, with respect
to 'GPL' compatibility (ignore the fact that 'GPL' includes the
non-OSI-Approved GPLv1). However, I think this is really not so, as
the 'compliance' in the 'GPL' case for section 4 purposes would not be
clear without section 3, paragraph 2. (?)