Revise 'Derived Work' definition to implement weak copyleft default rule.
The purpose of this change is best explained with an example.

If a Python file foo.py contains the line 'import bar', where bar is a
library licensed under copyleft-next, the mere inclusion of 'import
bar' without anything more is not in itself enough to make foo.py a
Derived Work of bar (regardless of whether it might otherwise fit the
definition of Derived Work).

However, if bar also contains some statement from the licensor like:

  Licensed under copyleft-next. Any code that imports bar, and which
  otherwise meets the definition of 'Derived Work' of bar in
  copyleft-next , shall be considered a Derived Work of bar even if
  such code is not distributed with bar.

then foo.py in itself falls within the copyleft scope of bar provided
that it meets the definition of 'Derived Work'.