Substantial rewrite of copyleft-related provisions.
Section 3 combines the pass-through, the basic copyleft requirement
for what I'm currently calling 'Modifications', and the prohibition
against imposition of further restrictions.

A new section 4 then collects 'Exceptions and Limitations': the later
version permission (which I feel needs some satisfactory legal
explanation built in?), the MPL 2.0-style GPL outbound licensing
permission, an awkward rewording of the inbound-compatibility
provision, and a new clause that says that 'Aggregates' (which will be
defined like 'Mere Aggregation') can be licensed arbitrarily restrictively as
long as the restrictions do not affect rights in the Covered Code.