Modified top definitions; added Covered Code definition; add 'lawful uses unaffected' provision.
The 'No Trademark License' clause is now placed in a separate section.

The 'Lawful Uses Unaffected' provision resembles a clause that was in
earlier releases of copyleft-next. It seems important to bring it back
since now we are avoiding explicit reliance on derivative work-like
legal constructs. For example, if You modify a file in 'My Code' in
such a way that copying no longer requires a license from Me, it still
will meet the technical definition of "Modified Code", but You can
treat it as outside the scope of the license.