Improvements to warranty/liability disclaimers.
This removes the MPL 2.0-style conspicuousness of sections 12 and 13.
New section 13a allows distributors to include additional disclaimers
or to include a copy of copyleft-next with, for example, all or part
of the contents of sections 12 and 13 capitalized.

I note that in three decades or so of free software there has been no
known instance in which lack of conspicuousness in warranty
disclaimers or limitation-of-liability provisions had any adverse
consequence. I further note that a number of FLOSS licenses carefully
drafted by lawyers outside the US have corresponding provisions with
no typographical effort to improve conspicuousness. These licenses do
not cover a large amount of software, but they cover packages found in
mainstream Linux distributions distributed in the United States as
well as other countries.