Lots of mostly formal changes; substantive change to license compatibility.
An appendix list of compatible licenses is introduced and is said to
be nonexhaustive and illustrative. The issue of conflicts between
copyleft-next and explicitly-compatible weak copyleft licenses (the
appendix lists MPL 2.0 and EPL) is addressed by saying that the 2c
copyleft condition can be ignored to the extent necessary to comply
with such weak copyleft licenses. It is also made clear that the the
license compatibility permission does not contract the scope of the
Corresponding Source requirement when Object Code is Distributed.