Eliminate 'Distribute'/'Distributor' as defined terms.
This was in a sense a partial legacy of GPLv3's 'internationalization'
feature. I am no longer convinced it is particularly useful. It is
true that now the license is clearly a creature of US copyright law
terminology, but the same may be said of so many FLOSS licenses which
see very wide use.