Remove upstream contribution licensing condition.
This was inspired by Apache License 2.0 section 5. I think the value
of including this is not worth the effort of getting it right.
copyleft-next projects will live without this provision in an
inbound=outbound world and can be encouraged to make use of the DCO or
similar contribution mechanisms. There is also an argument that
including this provision suggests that it is necessary.