Remove superflous “without more” in Derivative Work definition
Commit [0eb4829: Attempt at definition of 'Proprietary License'
changed the definition of the Derivative Work and in particular
the second exception from:

	(ii) a mere reproduction of My Code;

to

	(ii) a mere reproduction of My Code, without more

i.e. “, without more” has been added.  As far as I can tell, that
sentence simply does not parse with that addition and it seems
something more was meant to follow, but did not.  Remove the
superflous words bringing the previous wording back.