Delete pointless clarification of "Licensed Patents" definition.
This clarification is taken from GPLv3, where it was inserted to
mollify concerns of patent-holding companies overly worried about the
potential effects of GPLv3 on their portfolios. It is obvious from the
definition of Licensed Patents that it cannot extend to claims first
infringed by further modification of My Code.