Make various changes suggested by Adam Saunders.
'Pertinent' in notice preservation provision is removed; Adam argues
this limitation could encourage noncompliance.

Add statutory damages to LoL provision.

Clarify exception to the source URL requirement in cases where
Products include Corresponding Source.

Get rid of definition of legal entity in 'I'/'You' definition. Adam
argues: "This makes it easier and cheaper to bring corporate entities
into compliance with copyleft-next; enforcement through litigation
won't need to delve as deeply into proving how an the infringing
corporate body is governed." I'm not sure I agree with that, but it's
fairly clear now that the absence of a legal entity definition in
GPLv2 and GPLv3 has not itself had any effect on corporate adoption of
or compliance with those licenses.