Clarify license compatibility provisions
The MPL 2.0-influenced GPL compatibility provision of section 3
arguably is problematic because it can be read as clashing with the
GPL's copyleft requirement to license the "entire" modified work under
the GPL. I believe the conflict is illusory because one cannot grant a
copyright license covering what one does not hold copyright
on. Nevertheless, having looked at this provision with fresh eyes, I
can foresee it being criticized. The change makes the provision more
like a typical relicensing clause in licenses that aim at outbound GPL
compatibility.

The latter part of section 4 is meant to cover the inbound license
compatibility problem, but the existing version is not clearly
inapplicable to the situation where one is combining copyleft-next
code with GPL code. I believe the change made here is useful as it
does not cover the case where you would have noncompliance with the
other license (as for example in a GPL|copyleft-next combination
scenario, if you ignore the existence of the latter part of section
3).