Delete Richard Stallman quote
I recently learned of some horrifying statements Stallman made
concerning children with Down syndrome. I can no longer countenance
anything that could be seen as an indication that Stallman should be
seen as an authority figure or figure worthy of respect. The only
purpose inclusion of this quote served was to provide some
legitimization of copyleft-next for that subset of people inclined to
think that a 'fork' of the GPL (as copyleft-next originally was) was
of questionable legitimacy if not authorized by the Free Software
Foundation.