Reorganize draft; mostly non-substantive changes
This is a significant reworking of the current draft, but most of the
changes aren't really substantive. Definitions once again are pushed
to the end. The 'pass through' statement is where I think it ought to
be, in the initial license grant section. Again, an effort to deal
with implementation of copyleft through terminology. Hopefully this
approach is a little simpler. It relies on the definition of 'Covered
Work'. The 'Enhancement' term is gone.